LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/inspector/client-close-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/inspector/echo' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 57 - Socket is not connected)
Tests WebSockets connection closed by the client.


== Running test suite: WebSocket.ClientClose
-- Running test case: WebSocket.ClientClose
PASS: WebSocketResource should be created.
PASS: WebSocketResource should be a WebSocket type.
PASS: WebSocketResource should be in a Connecting state.
PASS: WebSocketResource previous state should be Connecting.
PASS: WebSocketResource state should change to Open.
PASS: WebSocketResource should be in a Open state.
PASS: WebSocketResource previous state should be Open.
PASS: WebSocketResource state should change to Closed.
PASS: WebSocketResource should be in a Closed state.
